Update main.c
Fix for possible overflow in cases of possibly very large values for pass_cnt overflowint  loop counter at main.c:302

+++ main.c
@@ -299,7 +299,7 @@ main(int argc, char **argv)
         exit(EXIT_FAILURE);
     }
 
-    for (int i = 0; i < pass_cnt; ++i) {
+    for (size_t i = 0; i < pass_cnt; ++i) {
         char *pass;
 
         /* We needn't add pass to tree since we free ASAP */
